requirements.txt
pandas
datasets
tqdm
matplotlib
google-genai
dotenv
markdown-it-py